+++ src/main/java/ladder/domain/generator/LadderStepsGenerator.java
@@ -1,0 +1,9 @@
+package ladder.domain.generator;
+
+import ladder.domain.ladder.LadderStep;
+
+import java.util.List;
+
+public interface LadderStepsGenerator {
+    List<LadderStep> generate();
+}
